Moving symbol table files to new crate
