+++ Cargo.toml
@@ -7,5 +7,5 @@ members=[
 	"lang-model",
 	"interpreter",
 	"interpreter/fuzz",
-	"ffi"]
+	"ffi", "symbol_table"]
