+++ symbol_table/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "symbol_table"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[features]
+fuzzing = []
+
+[dependencies]
+clap = { version = "4.5.16", features = ["derive"] }
+libc = "0.2.158"
+uname = "0.1.1"
+thiserror = "1.0.63"
+ffi = { path = "../ffi"}
+ref-cast = "1.0.23"
+# Note I am currently enabling all the derive more features for conveniences.
+# If compile times become an issue strip this back to just what is needed.
+derive_more = { version = "1", features = ["full"] }
+arbitrary = { version = "1.3.2", features = ["derive"] }
+
+[dev-dependencies]
+uuid = { version = "1.10.0", features = ["v4"] }
+rstest = "0.22.0"
+pretty_assertions = "1.4.0"
+rand = "0.8.5"
+mutants = "0.0.3"
+
+[build-dependencies]
+bindgen = "0.70.1"
+cc = "1.1.16"
+cbindgen = "0.27.0"
+glob = "0.3.1"
