+++ symbol_table/src/var_data/mod.rs
