+++ symbol_table/src/var_data/tests.rs
