+++ symbol_table/src/var_u.rs
